Extract the `DPhyp` join-order optimizer
